tests/requirements.txt
Appium-Python-Client==0.22
jsonpickle==0.9.3
macholib==1.5.1
pyparsing==2.0.1
pytest==2.9.1
selenium==2.53.1
tornado==4.3
websocket-client==0.37.0
pycurl==7.43.0
termcolor==1.1.0
